ai/fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
boto3==1.38.5
botocore==1.38.5
cachetools==5.5.2
cassidy==0.1.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
diffusers==0.33.1
distro==1.9.0
enumb==0.1.5
exceptiongroup==1.2.2
expo==0.1.2
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
MarkupSafe==2.1.5
mimetype==0.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
openai==1.76.2
packaging==25.0
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
zipp==3.21.0
inflect==7.5.0
einops==0.8.1
sentencepiece==0.2.0
kanjize==1.6.0
phonemizer==3.3.0
sudachipy==0.6.10
sudachidict_fullpython-multipart
